django-restframework/drf_lessions/requirements.txt
asgiref==3.4.1
awesome-slugify==1.6.5
certifi==2021.10.8
charset-normalizer==2.0.7
defusedxml==0.7.1
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-xml==2.0.0
djangorestframework-yaml==2.0.0
httpie==2.6.0
idna==3.3
Markdown==3.3.4
Pillow==8.4.0
Pygments==2.10.0
PySocks==1.7.1
pytz==2021.3
PyYAML==6.0
regex==2021.10.8
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.2
Unidecode==0.4.21
urllib3==1.26.7
